[rename]: Column 폴더 이름 변경 (#82)
components/feature/ProductItem/
- 기존: Column
- 변경: ColumnProductItem
